Added header guards to all header files Changed extension of script_replacement.hpp to .cpp
