+++ source/lua_bind_hash.hpp
@@ -1,3 +1,6 @@
+#ifndef LUA_BIND_HASH_H
+#define LUA_BIND_HASH_H
+
 #include <stdio.h>
 #include <string.h>
 #include <stdint.h>
@@ -94,3 +97,5 @@ uint64_t lua_bind_hash(const void* data_, size_t len) {
 uint64_t lua_bind_hash_str(const char* str) {
 	return lua_bind_hash(str, strlen(str));
 }
+
+#endif // LUA_BIND_HASH_H
